+++ src/components/ColorBarCanvas.tsx
@@ -26,12 +26,7 @@ import * as React from 'react';
 import { useEffect, useRef } from 'react';
 
 import i18n from '../i18n';
-import { ColorBar } from '../model/colorBar';
-
-import bgImageData from "./bg.png";
-
-const bgImage = new Image();
-bgImage.src = bgImageData;
+import { ColorBar, renderColorBar } from '../model/colorBar';
 
 
 interface ColorBarCanvasProps {
@@ -43,115 +38,34 @@ interface ColorBarCanvasProps {
 }
 
 export const ColorBarCanvas: React.FC<ColorBarCanvasProps> = (
-        {
-            colorBar,
-            opacity,
-            width,
-            height,
-            onOpenEditor
-        }
+    {
+        colorBar,
+        opacity,
+        width,
+        height,
+        onOpenEditor
+    }
 ) => {
     const canvasRef = useRef<HTMLCanvasElement | null>(null);
 
     useEffect(() => {
         const canvas = canvasRef.current;
-        if (canvas === null || !Boolean(colorBar.imageData)) {
-            return;
+        if (canvas !== null) {
+            renderColorBar(colorBar, opacity, canvas);
         }
-        const image = new Image();
-        image.onload = () => {
-            renderColorBar(colorBar, opacity, canvas, image);
-        };
-        image.src = `data:image/png;base64,${colorBar.imageData}`;
     }, [colorBar, opacity]);
 
     return <>
         {Boolean(colorBar.imageData) ? (
-                <canvas
-                        ref={canvasRef}
-                        width={width || 240}
-                        height={height || 24}
-                        onClick={onOpenEditor}
-                />
-        ) : (
-                <div>{i18n.get('Unknown color bar') + `: ${colorBar.baseName}`}</div>
+            <canvas
+                ref={canvasRef}
+                width={width || 240}
+                height={height || 24}
+                onClick={onOpenEditor}
+            />
+        ):(
+            <div>{i18n.get('Unknown color bar') + `: ${colorBar.baseName}`}</div>
         )}
     </>;
 }
-
-
-function renderColorBar(colorBar: ColorBar,
-                        opacity: number,
-                        canvas: HTMLCanvasElement,
-                        image: HTMLImageElement) {
-    const offscreenCanvas = document.createElement('canvas');
-    offscreenCanvas.width = image.width;
-    offscreenCanvas.height = image.height;
-    const ctx = offscreenCanvas.getContext("2d");
-    if (ctx === null) {
-        return;
-    }
-    ctx.drawImage(image, 0, 0);
-    const imageData = ctx.getImageData(
-            0, 0,
-            offscreenCanvas.width,
-            offscreenCanvas.height
-    );
-
-    let rgbaArray = imageData.data;
-
-    if (colorBar.isReversed) {
-        const reversedRgbaArray = new Uint8ClampedArray(rgbaArray.length);
-        for (let i = 0; i < rgbaArray.length; i += 4) {
-            const j = rgbaArray.length - i - 4;
-            reversedRgbaArray[j] = rgbaArray[i];
-            reversedRgbaArray[j + 1] = rgbaArray[i + 1];
-            reversedRgbaArray[j + 2] = rgbaArray[i + 2];
-            reversedRgbaArray[j + 3] = rgbaArray[i + 3];
-        }
-        rgbaArray = reversedRgbaArray;
-    }
-
-    if (colorBar.isAlpha) {
-        const factor = 256 * 2 / rgbaArray.length;
-        for (let i = 0; i < rgbaArray.length / 2; i += 4) {
-            rgbaArray[i + 3] = factor * i;
-        }
-    }
-
-    if (opacity < 1.0) {
-        for (let i = 0; i < rgbaArray.length; i += 4) {
-            rgbaArray[i + 3] *= opacity;
-        }
-    }
-
-    Promise.resolve(
-            createImageBitmap(
-                    new ImageData(rgbaArray,
-                            rgbaArray.length / 4,
-                            1)
-            )
-    ).then(imageBitmap => {
-        const ctx = canvas.getContext("2d");
-        if (ctx !== null) {
-            const pattern = ctx.createPattern(bgImage, "repeat");
-            if (pattern !== null) {
-                ctx.fillStyle = pattern;
-            } else {
-                ctx.fillStyle = "#ffffff";
-            }
-            ctx.fillRect(
-                    0, 0,
-                    canvas.width, canvas.height
-            );
-            ctx.drawImage(
-                    imageBitmap,
-                    0, 0,
-                    canvas.width, canvas.height
-            );
-        }
-    });
-}
-
-
