+++ src/model/colorBar.ts
@@ -22,6 +22,9 @@
  * SOFTWARE.
  */
 
+import bgImageData from "./bg.png";
+
+
 export interface ColorBars {
     groups: ColorBarGroup[];
     images: { [name: string]: string };
@@ -35,6 +38,10 @@ export interface ColorBarGroup {
 
 export const CB_ALPHA_SUFFIX = '_alpha';
 export const CB_REVERSE_SUFFIX = '_r';
+
+
+const BG_IMAGE = new Image();
+BG_IMAGE.src = bgImageData;
 
 
 export interface ColorBar {
@@ -73,3 +80,98 @@ export function formatColorBar(colorBar: ColorBar): string {
     }
     return name;
 }
+
+export function renderColorBar(colorBar: ColorBar,
+                               opacity: number,
+                               canvas: HTMLCanvasElement) {
+    loadColorBarImageData(colorBar, opacity).then(imageData => {
+        Promise.resolve(
+            createImageBitmap(imageData)
+        ).then(imageBitmap => {
+            const ctx = canvas.getContext("2d");
+            if (ctx !== null) {
+                const pattern = ctx.createPattern(BG_IMAGE, "repeat");
+                if (pattern !== null) {
+                    ctx.fillStyle = pattern;
+                } else {
+                    ctx.fillStyle = "#ffffff";
+                }
+                ctx.fillRect(
+                    0, 0,
+                    canvas.width, canvas.height
+                );
+                ctx.drawImage(
+                    imageBitmap,
+                    0, 0,
+                    canvas.width, canvas.height
+                );
+            }
+        });
+    });
+}
+
+export function loadColorBarImageData(colorBar: ColorBar, opacity: number): Promise<ImageData> {
+    return new Promise<ImageData>(
+        (resolve: (value?: ImageData | PromiseLike<ImageData>) => void, reject: (reason?: any) => void) => {
+            const image = new Image();
+            image.onload = () => {
+                const imageData = getColorBarImageData(colorBar, opacity, image);
+                if (imageData !== null) {
+                    resolve(imageData);
+                } else {
+                    reject("failed to retrieve 2d context");
+                }
+            };
+            image.src = `data:image/png;base64,${colorBar.imageData}`;
+        }
+    );
+}
+
+function getColorBarImageData(colorBar: ColorBar,
+                              opacity: number,
+                              image: HTMLImageElement): ImageData | null {
+    const offscreenCanvas = document.createElement('canvas');
+    offscreenCanvas.width = image.width;
+    offscreenCanvas.height = image.height;
+    const ctx = offscreenCanvas.getContext("2d");
+    if (ctx === null) {
+        return null;
+    }
+    ctx.drawImage(image, 0, 0);
+    const imageData = ctx.getImageData(
+        0, 0,
+        offscreenCanvas.width,
+        offscreenCanvas.height
+    );
+
+    let rgbaArray = imageData.data;
+
+    if (colorBar.isReversed) {
+        const reversedRgbaArray = new Uint8ClampedArray(rgbaArray.length);
+        for (let i = 0; i < rgbaArray.length; i += 4) {
+            const j = rgbaArray.length - i - 4;
+            reversedRgbaArray[j] = rgbaArray[i];
+            reversedRgbaArray[j + 1] = rgbaArray[i + 1];
+            reversedRgbaArray[j + 2] = rgbaArray[i + 2];
+            reversedRgbaArray[j + 3] = rgbaArray[i + 3];
+        }
+        rgbaArray = reversedRgbaArray;
+    }
+
+    if (colorBar.isAlpha) {
+        const factor = 256 * 2 / rgbaArray.length;
+        for (let i = 0; i < rgbaArray.length / 2; i += 4) {
+            rgbaArray[i + 3] = factor * i;
+        }
+    }
+
+    if (opacity < 1.0) {
+        for (let i = 0; i < rgbaArray.length; i += 4) {
+            rgbaArray[i + 3] *= opacity;
+        }
+    }
+
+    return new ImageData(rgbaArray, rgbaArray.length / 4, 1);
+}
+
+
